requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
requests
psycopg2-binary
python-dotenv
SQLAlchemy
gunicorn
azure-identity